+++ src/Application/Adapter/CachedNetworkingStatusInterface.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Proximum\Vimeet\Application\Adapter;
+
+
+interface CachedNetworkingStatusInterface
+{
+    public function isOnline(int $eventId, int $userId): bool;
+}