+++ src/Ui/Bundle/EventBundle/Resources/assets/js/components/Toast/NotificationToastManager.js
@@ -3,11 +3,14 @@ import NotificationToast from '../Toast/NotificationToast';
 class NotificationToastManager {
 
     constructor(element, container, currentUserConnection) {
+        this.listeners = null;
+
         if (element && container) {
             this.element = element;
             this.container = container;
             currentUserConnection.addListener(this.onNotificationReceived.bind(this));
             this.toasts = [];
+            this.listeners = [];
         }
     }
 
@@ -28,12 +31,37 @@ class NotificationToastManager {
                 toastToRemove.destroy();
             }
 
-            toast.element.querySelector('.close').addEventListener('click', ()=> {
+            const close = () => {
                 const index = this.toasts.indexOf(toast);
                 this.toasts.splice(index,1);
                 toast.destroy();
+            };
+
+            toast.element.querySelector('.close').addEventListener('click', close);
+
+            toast.element.querySelector('.message-link').addEventListener('click', () => {
+                this.listeners.forEach((callback) => callback(toast));
+                close();
             });
         }
+    }
+
+    addListener(callback) {
+        if (this.listeners === null) {
+            return;
+        }
+
+        this.listeners.push(callback);
+    }
+
+    removeListener(callback) {
+        if (this.listeners === null) {
+            return;
+        }
+
+        this.listeners = this.listeners.filter((item) => {
+            return item !== callback
+        })
     }
 }
 export default NotificationToastManager;